+++ README.md
@@ -1,15 +1,13 @@
-# venic
+# veno
 ```
-   ██▒   █▓▓█████  ███▄    █  ██▓ ▄████▄  
-  ▓██░   █▒▓█   ▀  ██ ▀█   █ ▓██▒▒██▀ ▀█  
-   ▓██  █▒░▒███   ▓██  ▀█ ██▒▒██▒▒▓█    ▄ 
-    ▒██ █░░▒▓█  ▄ ▓██▒  ▐▌██▒░██░▒▓▓▄ ▄██▒
-     ▒▀█░  ░▒████▒▒██░   ▓██░░██░▒ ▓███▀ ░
-     ░ ▐░  ░░ ▒░ ░░ ▒░   ▒ ▒ ░▓  ░ ░▒ ▒  ░
-     ░ ░░   ░ ░  ░░ ░░   ░ ▒░ ▒ ░  ░  ▒   
-       ░░     ░      ░   ░ ░  ▒ ░░        
-        ░     ░  ░         ░  ░  ░ ░      
-       ░                         ░        
+                 #               
+                 ##              
+ ###  ## ####### ###  ##  ###### 
+ ###  ## #       #### ##  ##   ##
+ ###  ## ####### #######  ##   ##
+  #####  ###     ### ###  ##   ##
+   ###   ####### ###  ##   ##### 
+                       #         
 ```
 Multi-purpose text/code editor meant for easy and vast expandability.
