+++ old/modules/keybindings/ctrl-G.py
