+++ old/modules/keybindings/ctrl-W.py
