+++ old/modules/magicBar.py
