+++ old/modules/syntaxHighlighting.py
