+++ old/tests/256colors2.pl
