+++ old/tests/curses-mouse-test.py
