+++ old/tests/curses-test-2.py
