+++ old/tests/curses-test.py
